Fixes for difficulty calculator related operations
- Object scales were not set properly
- Strain peak sorting mutated the original list
- `TrackAttributeSet` does not accept `double`s

+++ src/com/reco1l/management/game/TrackAttributeSet.java
@@ -164,7 +164,7 @@ public class TrackAttributeSet implements ModAcronyms {
             StandardDifficultyAttributes difficultyAttributes =
                     BeatmapDifficultyCalculator.calculateStandardDifficulty(beatmap);
 
-            get(TrackAttribute.STARS).opt(v -> difficultyAttributes.starRating);
+            get(TrackAttribute.STARS).opt(v -> (float) difficultyAttributes.starRating);
         });
     }
 }
